Split out state resolving logic from instance onboarding
