+++ enderchest/filesystem.py
@@ -1,9 +1,10 @@
 """Functionality for managing the EnderChest and shulker box config files and folders"""
+
 import os
 from pathlib import Path
 from typing import Iterable
 
-from .loggers import GATHER_LOGGER
+from .loggers import LOAD_LOGGER
 
 ENDER_CHEST_FOLDER_NAME = "EnderChest"
 
@@ -170,7 +171,7 @@ def shulker_box_configs(minecraft_root: Path) -> Iterable[Path]:
     This method does not check to make sure those config files are valid,
     just that they exist
     """
-    GATHER_LOGGER.debug(f"Searching for shulker configs within {minecraft_root}")
+    LOAD_LOGGER.debug(f"Searching for shulker configs within {minecraft_root}")
     return ender_chest_folder(minecraft_root).glob(f"*/{SHULKER_BOX_CONFIG_NAME}")
