+++ enderchest/loggers.py
@@ -1,9 +1,11 @@
 """Loggers for the various EnderChest actions"""
+
 import logging
 
 BREAK_LOGGER = logging.getLogger("enderchest.break")
 CRAFT_LOGGER = logging.getLogger("enderchest.craft")
 GATHER_LOGGER = logging.getLogger("enderchest.gather")
+LOAD_LOGGER = logging.getLogger("enderchest.load")
 PLACE_LOGGER = logging.getLogger("enderchest.place")
 SYNC_LOGGER = logging.getLogger("enderchest.sync")
